Make SupportsDelete to be simple mix-in of DSV2
