--- sql/catalyst/src/main/java/org/apache/spark/sql/sources/v2/SupportsMaintenance.java
@@ -1,35 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.spark.sql.sources.v2;
-
-import org.apache.spark.sql.sources.v2.maintain.Maintainer;
-import org.apache.spark.sql.sources.v2.maintain.MaintainerBuilder;
-import org.apache.spark.sql.util.CaseInsensitiveStringMap;
-
-/**
- * A mix-in interface of {@link Table}, to indicate that it can be maintained. This adds
- *  {@link #newMaintainerBuilder(CaseInsensitiveStringMap)} that is used to create a
- *  {@link Maintainer}
- */
-public interface SupportsMaintenance {
-
-  /**
-   * Returns an {@link MaintainerBuilder} which can be used to
-   */
-  MaintainerBuilder newMaintainerBuilder(CaseInsensitiveStringMap options);
-}
